+++ plugins/modules/cp_gaia_grub_password.py
@@ -21,55 +21,48 @@ from __future__ import (absolute_import, division, print_function)
 
 __metaclass__ = type
 
+
 DOCUMENTATION = """
-module: cp_gaia_system_group
-author: Duane Toler (@duanetoler)
+author: Ameer Asli (@chkp-ameera)
 description:
-- Change system group configuration.
-short_description: Change system group configuration.
-version_added: '8.0.0'
+- Sets GRUB password.
+module: cp_gaia_grub_password
+short_description: Sets GRUB password.
+version_added: '6.0.0'
 notes:
 - Supports C(check_mode).
+requirements:
+- supported starting from gaia_api >= 1.7
 options:
-  version:
-    description: Gaia API version for example 1.6.
-    required: False
-    type: str
-  state:
-    description: Ansible state which can be C(present) or C(absent).
-    required: False
-    type: str
-    default: present
-    choices: [present, absent]
-  name:
-    description: System group name.
-    required: True
-    type: str
-  gid:
-    description: Specifies a numeric group ID used to identify the group, duplicate GIDs are not allowed.
-    required: False
-    type: int
-  users:
-    description: List of users in the group.
-    required: False
-    type: list
-    elements: str
+    version:
+      description: Gaia API version for example 1.7.
+      required: False
+      type: str
+    password:
+        description: GRUB new password.
+        required: False
+        type: str
+    password_hash:
+        description: An encrypted representation of the password.
+        required: False
+        type: str
 """
+
 
 EXAMPLES = """
-- name: Add new system group
-  check_point.gaia.cp_gaia_system_group:
-    name: new_group
-    members:
-      - admin
+- name: Setting GRUB new password
+  check_point.gaia.cp_gaia_grub_password:
+    password: newpass
 """
 
+
 RETURN = """
-group:
-  description: The updated group details.
+grub_password:
+  description: The checkpoint object updated.
   returned: always.
   type: dict
 """
+
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.check_point.gaia.plugins.module_utils.checkpoint import chkp_api_call, checkpoint_argument_spec_for_all
@@ -78,17 +71,14 @@ from ansible_collections.check_point.gaia.plugins.module_utils.checkpoint import
 def main():
     # arguments for the module:
     fields = dict(
-        state=dict(type='str', default='present', choices=['present', 'absent']),
-        name=dict(type='str', required=True),
-        gid=dict(type='int', required=False),
-        users=dict(type='list', elements='str', required=False)
+        password=dict(type='str', no_log=True),
+        password_hash=dict(type='str', no_log=True)
     )
     fields.update(checkpoint_argument_spec_for_all)
     module = AnsibleModule(argument_spec=fields, supports_check_mode=True)
-    api_call_object = 'system-group'
-    show_params = ["name"]
+    api_call_object = 'grub-password'
 
-    res = chkp_api_call(module, api_call_object, True, show_params=show_params)
+    res = chkp_api_call(module, api_call_object, False)
     module.exit_json(**res)