+++ plugins/modules/cp_gaia_snmp_custom_trap_facts.py
@@ -21,67 +21,60 @@ from __future__ import (absolute_import, division, print_function)
 
 __metaclass__ = type
 
-
 DOCUMENTATION = """
+module: cp_gaia_snmp_custom_trap_facts
+author: Ameer Asli (@chkp-ameera)
 description:
-- Show the VSNext state.
-module: cp_gaia_vsnext_state_facts
-options:
-  version:
-    description: Gaia API version for example 1.8.
-    required: False
-    type: str
-short_description: Show the VSNext state.
+- Show the SNMP custom trap currently configured.
+short_description: Show SNMP custom traps.
 version_added: '6.0.0'
-author: Omer Hadad (@chkp-omerhad)
-
 notes:
 - Supports C(check_mode).
-requirements:
-- supported starting from gaia_api >= 1.8
-"""
-EXAMPLES = """
-- name: Show VSNext state
-  check_point.gaia.cp_gaia_vsnext_state_facts:
+options:
+  version:
+    description: Gaia API version for example 1.6.
+    required: False
+    type: str
+  name:
+    description: Custom trap name.
+    required: false
+    type: str
 
 """
+
+EXAMPLES = """
+- name: Show SNMP custom trap
+  check_point.gaia.cp_gaia_snmp_custom_trap_facts:
+
+- name: Show SNMP custom trap by specifying it's name
+  check_point.gaia.cp_gaia_snmp_custom_trap_facts:
+    name: custom_trap_name
+
+"""
+
 RETURN = """
 ansible_facts:
-    description: The VSNext state facts.
+    description: The checkpoint object facts.
     returned: always.
     type: dict
-    contains:
-        enabled:
-            description: The VSNext state.
-            returned: always
-            type: bool
-        session-virtual-system-id:
-            description: The Virtual System ID of the current Gaia API session.
-            returned: always
-            type: int
-        member-id:
-            description: The member on which the command was executed.
-            returned: On Scalable and Elastic XL platforms only.
-            type: sp-member-id
 """
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.check_point.gaia.plugins.module_utils.checkpoint import chkp_facts_api_call, checkpoint_argument_spec_for_all
 
 
-def run_module():
-    fields = dict()
+def main():
+    # arguments for the module:
+    fields = dict(
+        name=dict(type="str", required=False)
+    )
     fields.update(checkpoint_argument_spec_for_all)
     module = AnsibleModule(argument_spec=fields, supports_check_mode=True)
-    api_call_object = 'vsnext-state'
+    api_call_object = "snmp-custom-trap"
 
-    res = chkp_facts_api_call(module, api_call_object, False)
+    res = chkp_facts_api_call(module, api_call_object, True)
     module.exit_json(ansible_facts=res["ansible_facts"])
 
 
-def main():
-    run_module()
-
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()